Use Int32 for depth.

+++ include/rgbd/depth_encoder.hpp
@@ -9,6 +9,6 @@ class DepthEncoder
 public:
     virtual ~DepthEncoder() {}
     virtual DepthCodecType getCodecType() noexcept = 0;
-    virtual Bytes encode(gsl::span<const int16_t> depth_values, bool keyframe) noexcept = 0;
+    virtual Bytes encode(gsl::span<const int32_t> depth_values, bool keyframe) noexcept = 0;
 };
 }
