+++ include/rgbd/rvl_encoder.hpp
@@ -9,7 +9,7 @@ class RVLEncoder : public DepthEncoder
 public:
     RVLEncoder(int width, int height) noexcept;
     DepthCodecType getCodecType() noexcept;
-    Bytes encode(gsl::span<const int16_t> depth_values, bool keyframe) noexcept;
+    Bytes encode(gsl::span<const int32_t> depth_values, bool keyframe) noexcept;
 
 private:
     const int width_;
